requirements.txt
appdirs==1.4.0
click==6.7
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe==0.23
packaging==16.8
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
Werkzeug==0.11.15
pypng
